Added bitmap processing
Created the BitmapLoaderService.ts file which takes bitmaps (explicitly .BMP files), processes them, and ensures the data makes it to the database. Part of this functionality includes changes to www.ts, which allows the bitmaps to be processed and the data stored automatically on server startup so we can change the maps easily.